crane: Extract webservice creation from __init__.py.
