+++ app.py
@@ -1,2 +1,2 @@
-from crane import app
+from crane.webserver import app
 app.run(debug=True)
