+++ crane/registry.py
@@ -1,4 +1,4 @@
-from crane import app, db
+from webserver import app, db
 import requests
 from flask import jsonify, request
 import json
